add weather daily screen

+++ app/src/main/java/com/android/jetpack/compose/ntduc/weather/presentation/weather_home/WeatherHomeScreen.kt
@@ -21,7 +21,7 @@ fun WeatherHomeScreen(
     modifier: Modifier,
     onRequestLocationPermission: () -> Unit,
     onEnableGps: () -> Unit,
-    onClickDailyWeather: () -> Unit
+    onClickDailyWeather: (Int) -> Unit
 ) {
     LaunchedEffect(key1 = Unit) {
         viewModel.loadWeatherInfo()
